+++ src/ui/MusicExplorerView.h
@@ -29,6 +29,7 @@ public:
     void setArtworkCache(ArtworkCache *cache);
     void setAlbums(const QVector<Album> &albums);
     void setTrackProvider(std::function<QVector<Track>(const Album &album)> provider);
+    void setTrackFlagResolver(std::function<bool(const Track &, const QString &)> resolver);
     void setQueueIsPlaylistSourced(bool sourced);
     void refreshExpandedTracks();
     void applyAlbumGridViewSettingsJson(const QString &json);
@@ -74,6 +75,7 @@ signals:
     void albumPlayNextTemporaryRequested(const QString &albumTitle);
     void albumAddToQueueTemporaryRequested(const QString &albumTitle);
     void albumAddToPlaylistRequested(const QStringList &albumTitles);
+    void albumStartRadioRequested(const QString &albumTitle);
     void albumRatingChanged(const QString &albumArtistName, const QString &albumTitle, int rating0To100);
     void trackActivated(const Track &track);
     void trackPlayNextRequested(const QVector<Track> &tracks);
@@ -85,6 +87,7 @@ signals:
     void propertiesRequested(const Track &track);
     void startRadioRequested(const Track &track);
     void startArtistRadioRequested(const QString &artistName);
+    void trackFlagChanged(const Track &track, const QString &flag, bool on);
     void trackRatingChanged(const Track &track, int rating0To100);
     void albumGridViewSettingsChanged();
     void trackTableViewSettingsChanged();